Refactor GoldSrc Fileloader
