--- include/snd_loader.hpp
@@ -1,10 +1,0 @@
-#pragma once
-#include <iostream>
-
-#include "alure2.h"
-
-class GoldSrcFileFactory final : public alure::FileIOFactory
-{
-public:
-  alure::UniquePtr<std::istream> openFile(const alure::String &name) noexcept override;
-};
